requirements_pyfeat.txt
opencv-python==4.8.1.78
numpy==1.24.3
py-feat==0.6.0
torch==2.0.1
torchvision==0.15.2
